feat: expandir los dartos del analisis, mostrar las graficas. base para seccion de chat ai

+++ lib/domain/models/analysis/data_average_all.dart
@@ -24,7 +24,7 @@ class DataAverageAll extends Data {
 
 class ResultAverage {
   double? average;
-  int? max;
+  double? max;
   double? min;
   String? sensor;
